requirements.txt
asyncio
tabulate
python-dateutil
spacy
Pillow
Flask
pyperclip